Fix BUTTON link message
Follow-up to a65d7cd, which failed for the main menu. That's because current_menu is -1 for the main menu.

Since current menu is either -1 or points to a MENU entry, it's safe to add 1 to it and bring it to a correct range.

While on it, remove redundant `if`.

Per bug report by RiderDavis.

Fixes #95.